Add files via upload
last moment changes, updating mine and Yonatans changes due to conflicts made during code with me in pycharm.

+++ Ex1/Project/Comparisons.py
@@ -97,7 +97,7 @@ def get_speed(elev: elevator):
     :param elev: represents an elevator
     :return: speed of current elevator
     """
-    return int(elev.speed)
+    return float(elev.speed)
 
 
 def get_floor_time(elev: elevator):
